% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{brain_viz}
\alias{brain_viz}
\title{Brain Visual}
\usage{
brain_viz(obj, jitter_val = 0.04, side = TRUE, regs = NULL)
}
\arguments{
\item{obj}{from `get_connectivity()`}

\item{jitter_val}{control how far away the overlapping lines are (default is .04)}

\item{side}{Visualize the side angle of the brain? Default is TRUE (and currently the only option)}
}
\description{
Provides a `ggplot2` figure of the connections of regions of the brain.
}
