% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{calc_diff}
\alias{calc_diff}
\title{Calculate Difference of Effect Sizes for Brain Viz}
\usage{
calc_diff(obj1, obj2, pattern = "lag|time")
}
\arguments{
\item{obj1}{from `get_connectivity()`}

\item{obj2}{from `get_connectivity()`}

\item{pattern}{the regular expression to grab the variables not to be included (e.g., the lag or time variables)}
}
\description{
Calculate Difference of Effect Sizes for Brain Viz
}
