% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{calc_ratio}
\alias{calc_ratio}
\title{Calculate Ratio for Brain Viz}
\usage{
calc_ratio(obj1, obj2)
}
\arguments{
\item{obj1}{from `get_connectivity()`}

\item{obj2}{from `get_connectivity()`}
}
\description{
Calculate Ratio for Brain Viz
}
