% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_seq.R
\name{calculate_seq}
\alias{calculate_seq}
\title{calculate_seq}
\usage{
calculate_seq(
  adj_mat,
  tuning_in = 0.5,
  tuning_out = 0.5,
  alpha = 1,
  seeds = 10,
  memory,
  given_seq
)
}
\arguments{
\item{tuning_in}{weight of the in-connect when evaluating the connectivity. float in \link{0,1}}

\item{tuning_out}{weight of the out-connect when evaluating the connectivity. float in \link{0,1}}

\item{alpha}{smooth parameter which reflects how much the connectivity rely on the temp result. float in \link{0,1}}

\item{memory}{memory length, which decides how many recent nodes are taken into consideration when calculating the connectivity. int}

\item{given_seq}{given seq generated by other criterion}

\item{adjmat}{adjacency matrix of a network which should have identical colnames and rownames}

\item{seed}{random seed used to initialize the first node in Node seq}

\item{jump}{possibility of choosing a random node regardless of the connectivity when generating a new node in node seq. float in \link{0,1}}

\item{length_seq}{length of the connectivity seq and node seq}
}
\value{
a list contenting node seq and connectivity seq
}
\description{
function to generate node seq and the corresponding connectivity seq on the given network according to a given seq
}
