% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_based_on_scores.R
\name{cd_based_on_score}
\alias{cd_based_on_score}
\title{cd_based_on_score}
\usage{
cd_based_on_score(adj_mat, seeds = 1)
}
\arguments{
\item{adj_mat}{adjacency matrix of the target undirected and weighted network}
}
\value{
partition list and community labels
}
\description{
iteratively divide the network into 2 section--communities, until the target function won't increase. Need function connectivity_adj_mat and cut_into_2 in advance
}
