% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_method.R
\name{compare_method}
\alias{compare_method}
\title{compare_method}
\usage{
compare_method(adj_mat, truth)
}
\arguments{
\item{adj_mat}{adjacency matrix of the target undirected and weighted network}

\item{truth}{partition of the node compared with}
}
\value{
NMI and ARI of the comparation result with "infomap","fastgreedy","walktrap","louvain" and "leidun"
}
\description{
compare community detection algorithm with the given partition of nodes using "infomap","fastgreedy","walktrap","louvain" and "leidun"
The comparation is performed on undirected and weighted network
}
