% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_adj_mat.R
\name{connectivity_adj_mat}
\alias{connectivity_adj_mat}
\title{connectivity_adj_mat}
\usage{
connectivity_adj_mat(adj_mat, self_loop = T)
}
\arguments{
\item{adj_mat}{adjacency matrix of the target directed and weighted network}

\item{self_loop}{whether the NULL model consider self loop}
}
\value{
target function matrix like modularity matrix
}
\description{
target function reveal the hierarchy structure
}
