% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_cut_thres.R
\name{connectivity_cut_thres}
\alias{connectivity_cut_thres}
\title{connectivity_cut_thres}
\usage{
connectivity_cut_thres(
  adj_mat,
  tuning_in = 0.5,
  tuning_out = 0.5,
  threshold = 0.75,
  alpha = 1,
  seeds = 1,
  memory
)
}
\arguments{
\item{tuning_in}{weight of the in-connect when evaluating the connectivity. float in \link{0,1}}

\item{tuning_out}{weight of the out-connect when evaluating the connectivity. float in \link{0,1}}

\item{alpha}{smooth parameter which reflects how much the connectivity rely on the temp result. float in \link{0,1}}

\item{memory}{memory length, which decides how many recent nodes are taken into consideration when calculating the connectivity. int}

\item{adjmat}{adjacency matrix of a network which should have identical colnames and rownames}

\item{seed}{random seed used to initialize the first node in Node seq}
}
\value{
membership: node_seq and the community label of each node in order of the connectivity seq

connect_record: connectivity seq

cluster: nodes and their community label in order of the truth community partition
}
\description{
function to generate node seq and the corresponding connectivity seq on the given network.
In this algorithm the connectivity seq will be cut according to the given threshold.
the cut of the connectivity will return a partition of node seq accordingly which served as community detection result.
}
