% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_projection2d.R
\name{connectivity_projection2d}
\alias{connectivity_projection2d}
\title{connectivity_projection2d}
\usage{
connectivity_projection2d(
  adj_mat,
  tuning_in = 0.5,
  tuning_out = 0.5,
  alpha = 0.3,
  seeds = 10,
  memory = 60
)
}
\arguments{
\item{tuning_in}{weight of the in-connect when evaluating the connectivity. float in \link{0,1}}

\item{tuning_out}{weight of the out-connect when evaluating the connectivity. float in \link{0,1}}

\item{alpha}{smooth parameter which reflects how much the connectivity rely on the temp result. float in \link{0,1}}

\item{memory}{memory length, which decides how many recent nodes are taken into consideration when calculating the connectivity. int}

\item{adjmat}{adjacency matrix of a network which should have identical colnames and rownames}

\item{seed}{random seed used to initialize the first node in Node seq}
}
\value{
node_seq: 2 node sequences generated
connectivity_seq: 2 connectivity sequences generated

coordinate: 2D coordinate generated
dip_result: result of the dip test on the 2D projection. A significant result means significant community structures
}
\description{
function to combine 2 connect seq to construct a 2D-projection of each node which retains the community structure.
in addition, we perform a dip test based on the projection to test whether there are significant cluster structure in the graph
}
