% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_seq.R
\name{connectivity_seq}
\alias{connectivity_seq}
\title{connectivity_seq}
\usage{
connectivity_seq(
  adj_mat,
  tuning_in = 0.5,
  tuning_out = 0.5,
  alpha = 1,
  seeds = 10,
  memory,
  sample_memory,
  jump = 0,
  length_seq
)
}
\arguments{
\item{tuning_in}{weight of the in-connect when evaluating the connectivity. float in \link{0,1}}

\item{tuning_out}{weight of the out-connect when evaluating the connectivity. float in \link{0,1}}

\item{alpha}{smooth parameter which reflects how much the connectivity rely on the temp result. float in \link{0,1}}

\item{memory}{memory length, which decides how many recent nodes are taken into consideration when calculating the connectivity. int}

\item{sample_memory}{denote how recent of those nodes in node seq won't be the candidate of the new node. int}

\item{jump}{possibility of choosing a random node regardless of the connectivity when generating a new node in node seq. float in \link{0,1}}

\item{length_seq}{length of the connectivity seq and node seq}

\item{adjmat}{adjacency matrix of a network which should have identical colnames and rownames}

\item{seed}{random seed used to initialize the first node in Node seq}
}
\value{
a list contenting node seq and connectivity seq
}
\description{
function to generate node seq and the corresponding connectivity seq on the given network
}
