% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_into_2.R
\name{cut_into_2}
\alias{cut_into_2}
\title{cut_into_2}
\usage{
cut_into_2(adj_mat_all, adj_mat, target_node, min_scale = 3, temp_seq_overview)
}
\arguments{
\item{adj_mat_all}{adjacency matrix of the target directed and weighted network}

\item{adj_mat}{local adjacency matrix of the target directed and weighted network}

\item{target_node}{node list}

\item{min_scale}{min scale to be cut}

\item{temp_seq_overview}{results returned by the connectivity_seq function}
}
\value{
two node lists and the corresponding adjacency matrix
}
\description{
cut the network into 2 sub-graph while taking the refine process into consideration.
}
