% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{effectsize_comp_viz}
\alias{effectsize_comp_viz}
\title{Effect Size Comparison Visual}
\usage{
effectsize_comp_viz(obj1, obj2)
}
\arguments{
\item{obj1}{from `get_connectivity()` with a grouping variable called `group`}

\item{obj2}{from `get_connectivity()` with a grouping variable called `group`}
}
\description{
Provides a `ggplot2` line graph of the connections of regions of the brain.
}
