% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{effectsize_viz}
\alias{effectsize_viz}
\title{Effect Size Visual}
\usage{
effectsize_viz(obj)
}
\arguments{
\item{obj}{from `get_connectivity()`}
}
\description{
Provides a `ggplot2` line graph of the connections of regions of the brain.
}
