% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_G_BR.R
\name{gererate_G_BR}
\alias{gererate_G_BR}
\title{gererate_G_BR}
\usage{
gererate_G_BR(trade, time_wanted)
}
\arguments{
\item{trade}{br_y or br_q}

\item{time_wanted}{time needed}
}
\value{
igraph object G and adjacency matrix
}
\description{
generate igraph object and adjacency matrix from BR_data
}
