% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connections.R
\name{get_connectivity}
\alias{get_connectivity}
\title{Get Brain Connectivity Measures}
\usage{
get_connectivity(obj, ..., lags)
}
\arguments{
\item{lags}{the number of lags to try}
}
\description{
Provides measures of connectivity between brain regions from a `xxxx` object.
}
