% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{ggplot.connectivity}
\alias{ggplot.connectivity}
\title{Brain Visual}
\usage{
ggplot.connectivity()
}
\arguments{
\item{...}{}
}
\description{
Provides a `ggplot2` figure of the connections of regions of the brain.
}
