% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_structure.R
\name{hierarchay_community}
\alias{hierarchay_community}
\title{hierarchy_community}
\usage{
hierarchay_community(adj_rely)
}
\arguments{
\item{adj_rely}{targrt function matrix}
}
\value{
combine seq and combine score
}
\description{
hierarchy bottom-up algorithm using target function
}
