% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{import_nirs}
\alias{import_nirs}
\title{Import Data From NIRS Output}
\usage{
import_nirs(path = "", ...)
}
\arguments{
\item{path}{the path to the main folder}

\item{...}{the channels (a named list of numbers)}
}
\description{
Import the various CSV files necessary to conduct connectivity analysis with some quick
cleaning, including getting the means of the regions based on the region variable
}
