% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity.R
\name{plot_connectivity}
\alias{plot_connectivity}
\title{plot_connectivity}
\usage{
plot_connectivity(node_seq, connectivity_seq, community)
}
\arguments{
\item{node_seq}{node seq generated by the connectivity analysis}

\item{connectivity_seq}{connectivity seq smoothed according to alpha}

\item{community}{community in order of the nodes' labels}
}
\value{
visualization of the connnectivity and the community partition
}
\description{
function to plot the connnectivity and the community partition
}
