% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity2d.R
\name{plot_connectivity2d}
\alias{plot_connectivity2d}
\title{plot_connectivity2d}
\usage{
plot_connectivity2d(coordinate, community, col_given = NULL)
}
\arguments{
\item{coordinate}{node seq generated by the connectivity analysis}

\item{community}{community in order of the nodes' labels. should be int vector}
}
\value{
visualization of the 2d-connnectivity and the community partition
}
\description{
function to plot the connnectivity 2D projection according to the given community partition
}
