% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_function.R
\name{plot_data_function}
\alias{plot_data_function}
\title{plot_data_function}
\usage{
plot_data_function(
  temp_seq,
  temp_connect_record,
  temp_community_df,
  seed = 1,
  col_given = NULL
)
}
\arguments{
\item{temp_seq}{node seq needed}

\item{temp_connect_record}{connectivity seq needed}

\item{temp_community_df}{data.frame composed of node labels and corresponding community labels}
}
\value{
data.frame composed of 4 columns which contents node, connect_record, community and color
}
\description{
generating data for visualization according to the given partition of the nodes in the connectivity analysis
}
