% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions_side}
\alias{regions_side}
\title{Regions Side default}
\format{A tibble with 8 rows and 3 variables:
\describe{
  \item{x}{location on the x axis}
  \item{y}{location on the y axis}
  \item{region}{the region name}
}}
\usage{
regions_side
}
\description{
The default list of regions and location, based on the side-view default brain diagram.
}
\keyword{datasets}
