% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculate.R
\name{score_calculate}
\alias{score_calculate}
\title{score_calculate}
\usage{
score_calculate(adj_rely, partition)
}
\arguments{
\item{adj_rely}{score function}

\item{partition}{node partition list}
}
\value{
score calculated according to the partition and
}
\description{
calculate the target function based on score matrix and partition
}
