% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_connectivity.R
\name{test_connectivity}
\alias{test_connectivity}
\title{parallel community affiliation test for nodes}
\usage{
test_connectivity(
  adj_mat,
  N = NULL,
  aplha = 0.7,
  num_cores = 8,
  m = 100,
  memory = 50,
  tuning_in = 1,
  tuning_out = 0
)
}
\arguments{
\item{adj_mat}{adjacency matrix of the graph which should be featured with colnames as well as rownames}

\item{N}{default to be number of nodes}

\item{aplha}{first order sliding smoothing parameter of the depedency sequence, default to be 0.7}

\item{num_cores}{parallel computation cores used,default to be 8}

\item{m}{number of sequences sampled from the network}

\item{memory}{sample memory, default to be 50}

\item{tuning_in}{importance of in-edges, default to be 1 which only make differences when the graph is directed}

\item{tuning_out}{importance of out-edges, default to be 1 which only make differences when the graph is directed}
}
\value{
test_stat:test statistics,p_value:test p values,test_record:all sampling results
}
\description{
parallel community affiliation test for nodes
}
