% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_test_function.R
\name{wavelet_test_function}
\alias{wavelet_test_function}
\title{wavelet_test_function}
\usage{
wavelet_test_function(
  adj_mat,
  memory = 60,
  m = 100,
  tuning_in = 0,
  tuning_out = 1,
  alpha = 0.4
)
}
\arguments{
\item{adj_mat}{adjacency matrix, N columns}

\item{memory}{dependence memory, default 60}

\item{m}{node seq generated by the connectivity analysis}

\item{tuning_in}{tuning_in,default 0}

\item{tuning_out}{tuning_out default 1}

\item{alpha}{smooth parameter for dependency value sequence}
}
\value{
wavelet_test_long result for G and random G，matrix recording p-value and test statistics
}
\description{
function to test community scale using wavelet analysis
}
