% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_test_long_rg.R
\name{wavelet_test_long_rg}
\alias{wavelet_test_long_rg}
\title{wavelet_test_long_rg}
\usage{
wavelet_test_long_rg(
  adj_mat,
  tuning_in = 0,
  tuning_out = 1,
  length_seq,
  memory = 60,
  sample_memory,
  m = 100,
  alpha = 1
)
}
\arguments{
\item{adj_mat}{adjacency matrix, N columns}

\item{tuning_in}{tuning_in,default 0}

\item{tuning_out}{tuning_out default 1}

\item{length_seq}{length of each seq}

\item{memory}{dependence memory, default 60}

\item{sample_memory}{sample_memory, default N}

\item{m}{node seq generated by the connectivity analysis}
}
\value{
power_all: power of different frequency at each node;node_all:node sequences in m samples; seq_all:dependence sequences in m samples; local_sum_mean:mean power at each node;local_sum_sd: sd of the power at each node;period: period of the wavelet basis
}
\description{
function to test community scale using wavelet analysis
}
