% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_viz.R
\name{brain_viz}
\alias{brain_viz}
\title{Brain Visual}
\usage{
brain_viz(obj, jitter_val = 0.04, view = "side", image = NULL,
  regs = NULL, diff = 0)
}
\arguments{
\item{obj}{from `get_connectivity()`}

\item{jitter_val}{control how far away the overlapping lines are (default is .04)}

\item{view}{the view of the brain diagram (available options are "side", "top", "right", "left"). If the image argument is given a path, then that image will be used instead of the built-in ones.}

\item{image}{If NULL then the default images (based on `view`); otherwise the path to the figure should be supplied here}

\item{regs}{Alternate locations for the regions of interest (needs to have x, y, and region as the variables)}

\item{diff}{For comparison brain viz's, this is the difference between the largest effect size of one to the other (current sample - comparison sample). Adjusts the size of the arrows to be more comparable across samples.}
}
\description{
Provides a `ggplot2` figure of the connections of regions of the brain.
}
