% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connections.R
\name{get_connectivity}
\alias{get_connectivity}
\title{Get Brain Connectivity Measures}
\usage{
get_connectivity(data, group = NULL, covariates = NULL, ...)
}
\arguments{
\item{data}{object from a nested tibble returned from `import_nirs()`}

\item{covariates}{variables to include in the model in addition to the regions of interest and the random effects}

\item{...}{arguments passed to `lme4::lmer()`}
}
\description{
Provides measures of connectivity between brain regions from a nested tibble returned from `import_nirs()`. It
fits the data to a multilevel model with a specified number of lags. It returns a tibble with effect size measures
to and from each region.
}
