% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{import_nirs}
\alias{import_nirs}
\title{Import Data From NIRS Output}
\usage{
import_nirs(path = "", ..., num_channels = 24, sides = FALSE,
  ignore_folders = NULL)
}
\arguments{
\item{path}{the path to the main folder}

\item{...}{the channels (a named list of numbers)}

\item{num_channels}{the number of channels for each of the probes}

\item{sides}{should the channel means be split into sides (based on the probe)}

\item{ignore_folders}{folders to ignore in the directory in regular expression form, e.g. "first|random" would remove any folders that have "first" or "random" in their names}
}
\description{
Import the various CSV files necessary to conduct connectivity analysis with some quick
cleaning, including getting the means of the regions based on the region variable
}
